get user from auth when post job

+++ src/main/java/com/example/gptunittestwriterjava/controller/UserController.java
@@ -15,7 +15,7 @@ public class UserController {
     private UserService userService;
 
     @GetMapping("/{id}")
-    public User getUser(@PathVariable Integer id) {
+    public User getUser(@PathVariable Long id) {
         return userService.getUser(id);
     }
